move finish options, logger, etc.
